Back Office terminé, pagination insérée

+++ src/Metinet/XtremQUIZZBundle/Entity/Quizz.php
@@ -42,7 +42,7 @@ class Quizz
     {
         return null === $this->picture
             ? null
-            : $this->getUploadRootDir().'/'.$this->id.'.'.$this->picture;
+            : $this->getUploadRootDir().'/'.$this->picture;
     }
     
     public function getWebPath()
@@ -651,7 +651,7 @@ class Quizz
         $this->getFile()->move($this->getUploadRootDir(), $this->picture);
 
         // check if we have an old image
-        if (isset($this->temp)) {
+        if (isset($this->temp) && file_exists($this->temp)) {
             // delete the old image
             unlink($this->temp);
             // clear the temp image path
@@ -671,6 +671,7 @@ class Quizz
      */
     public function storeFilenameForRemove()
     {
+        echo $this->getAbsolutePath();
         $this->temp = $this->getAbsolutePath();
     }
 
@@ -679,7 +680,7 @@ class Quizz
      */
     public function removeUpload()
     {
-        if (isset($this->temp)) {
+        if (isset($this->temp) && file_exists($this->temp)) {
             unlink($this->temp);
         }
     }
